Cleans up ResourceBuilder
Many methods and switch statements in ResourceBuilder were not in any particular order. Now switch statements are in alphabetical order, and the following methods are in the order defined in buildObjectsFromRawData, grouped by scope: create/update methods are nearby, and if some extra methods are needed (such as buildFile) they immediately follow the method which calls them.